moved `evalTerm` to `Evaluate.hs`.
