+++ Plutarch/Internal.hs
@@ -1,4 +1,3 @@
-{-# LANGUAGE ImpredicativeTypes #-}
 {-# LANGUAGE RoleAnnotations #-}
 
 module Plutarch.Internal (
@@ -35,7 +34,6 @@ module Plutarch.Internal (
   TracingMode (..),
   pgetConfig,
   TermMonad (..),
-  evalTerm,
 ) where
 
 import Control.Monad.Reader (ReaderT (ReaderT), ask, runReaderT)
@@ -54,11 +52,10 @@ import Data.Text (Text)
 import qualified Flat.Run as F
 import GHC.Stack (HasCallStack, callStack, prettyCallStack)
 import GHC.Word (Word64)
-import Plutarch.Evaluate (EvalError, evalScript, evalScriptHuge)
+import Plutarch.Internal.Evaluate (evalScript)
 import PlutusCore (Some (Some), ValueOf (ValueOf))
 import qualified PlutusCore as PLC
 import PlutusCore.DeBruijn (DeBruijn (DeBruijn), Index (Index))
-import PlutusCore.Evaluation.Machine.ExBudget (ExBudget)
 import PlutusLedgerApi.V1.Scripts (Script (Script))
 import qualified UntypedPlutusCore as UPLC
 
@@ -474,19 +471,3 @@ compile config t = case asClosedRawTerm t of
 
 hashTerm :: Config -> ClosedTerm a -> Either Text Dig
 hashTerm config t = hashRawTerm . getTerm <$> runReaderT (runTermMonad $ asRawTerm t 0) config
-
--- | Compile and evaluate term.
-evalTerm ::
-  Config ->
-  ClosedTerm a ->
-  Either Text (Either EvalError (ClosedTerm a), ExBudget, [Text])
-evalTerm config term = 
-  case compile config term of
-    Right script ->
-      let (s, b, t) = evalScriptHuge script
-       in Right (fromScript <$> s, b, t)
-    Left a -> Left a
-  where
-    fromScript :: Script -> ClosedTerm a
-    fromScript (Script script) =
-        Term $ const $ pure $ TermResult (RCompiled $ UPLC._progTerm $ script) []
